+++ dask/array/__init__.py
@@ -1,9 +1,9 @@
 from __future__ import absolute_import, division, print_function
 
 from ..utils import ignoring
-from .core import (Array, concatenate, stack, from_array, store, map_blocks,
-                   atop, to_hdf5, to_npy_stack, from_npy_stack, from_delayed,
-                   asarray, asanyarray, broadcast_to)
+from .core import (Array, block, concatenate, stack, from_array, store,
+                   map_blocks, atop, to_hdf5, to_npy_stack, from_npy_stack,
+                   from_delayed, asarray, asanyarray, broadcast_to)
 from .routines import (take, choose, argwhere, where, coarsen, insert,
                        ravel, roll, unique, squeeze, topk, ptp, diff, ediff1d,
                        bincount, digitize, histogram, cov, array, dstack,
@@ -11,7 +11,8 @@ from .routines import (take, choose, argwhere, where, coarsen, insert,
                        flatnonzero, nonzero, around, isnull, notnull, isclose,
                        allclose, corrcoef, swapaxes, tensordot, transpose, dot,
                        vdot, matmul, apply_along_axis, apply_over_axes,
-                       result_type, atleast_1d, atleast_2d, atleast_3d)
+                       result_type, atleast_1d, atleast_2d, atleast_3d,
+                       flip, flipud, fliplr)
 from .reshape import reshape
 from .ufunc import (add, subtract, multiply, divide, logaddexp, logaddexp2,
         true_divide, floor_divide, negative, power, remainder, mod, conj, exp,
@@ -24,6 +25,11 @@ from .ufunc import (add, subtract, multiply, divide, logaddexp, logaddexp2,
         copysign, nextafter, spacing, ldexp, fmod, floor, ceil, trunc, degrees,
         radians, rint, fix, angle, real, imag, clip, fabs, sign, absolute,
         i0, sinc, nan_to_num, frexp, modf, divide)
+try:
+    from .ufunc import float_power
+except ImportError:
+    # Absent for NumPy versions prior to 1.12.
+    pass
 from .reductions import (sum, prod, mean, std, var, any, all, min, max, vnorm,
                          moment,
                          argmin, argmax,